+++ python/taichi/ui/__init__.py
@@ -1,2 +1,3 @@
 from .gui import *
+from .imgui import *
 from .ui import *
