+++ python/taichi/ui/window.py
@@ -5,7 +5,7 @@ from taichi.lang.impl import default_cfg
 
 from .canvas import Canvas
 from .constants import PRESS, RELEASE
-from .gui import Gui
+from .imgui import Gui
 
 
 class Window(_ti_core.PyWindow):
